duman/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "happycake"
version = "0.1.0"
description = "HappyCake US — AI Sales & Operations OS (Steppe Business Club hackathon)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-telegram-bot>=21.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "jupyter>=1.1",
    "nbconvert>=7.16",
    "matplotlib>=3.9",
    "rich>=13.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["happycake*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S"]
ignore = ["S101", "S603", "S607"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers"
markers = [
    "scenario: YAML scenario tests",
    "unit: fast unit tests",
    "operator: telegram callback tests",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
